requirements.txt
feedparser
trafilatura
google-generativeai
jinja2
markdown